Propagate directory move changes to fact PR.
